Collections page converted to RSC, its form now a client component
